Note that CompiledBlock is still used even though it has been removed from master, because it is used when bruteforcing culling codes. That may be a bad thing.

Conflicts:
	Data/Sys/GameSettings/GALE01r2.ini
	Source/Core/Core/PowerPC/Interpreter/Interpreter_Branch.cpp
	Source/Core/Core/PowerPC/Jit64/Jit.cpp
	Source/Core/DolphinWX/Frame.cpp
	Source/Core/VideoBackends/D3D/Render.cpp
	Source/Core/VideoBackends/OGL/GLInterface/GLX.h
	Source/Core/VideoCommon/OpcodeDecoding.cpp

+++ Source/Core/Core/HW/EXI_DeviceIPL.h
@@ -20,7 +20,7 @@ public:
 	void DoState(PointerWrap &p) override;
 
 	static u32 GetGCTime();
-	static u32 NetPlay_GetGCTime();
+	static u64 NetPlay_GetGCTime();
 
 	static void Descrambler(u8* data, u32 size);
